Extract `Cell_map_misc_tests`
