+++ reactive/src/commonTest/kotlin/dev/toolkt/reactive/cell/Cell_map_combo_tests.kt
@@ -164,43 +164,4 @@ class Cell_map_combo_tests {
             )
         }
     }
-
-    private fun test_deactivation(
-        sourceCellFactory: DynamicCellFactory,
-        verificationStrategy: CellVerificationStrategy.Active,
-    ) {
-        val doTrigger = EmitterEventStream<Unit>()
-
-        val sourceCell = sourceCellFactory.createDynamicExternally(
-            initialValue = 10,
-            doUpdate = doTrigger.map { 11 },
-        )
-
-        val mapCell = sourceCell.map { it.toString() }
-
-        verificationStrategy.verifyDeactivation(
-            subjectCell = mapCell,
-            doTrigger = doTrigger,
-        )
-    }
-
-    private fun test_deactivation(
-        verificationStrategy: CellVerificationStrategy.Active,
-    ) {
-        DynamicCellFactory.values.forEach { sourceCellFactory ->
-            test_deactivation(
-                sourceCellFactory = sourceCellFactory,
-                verificationStrategy = verificationStrategy,
-            )
-        }
-    }
-
-    @Test
-    fun test_deactivation() {
-        CellVerificationStrategy.Active.values.forEach { verificationStrategy ->
-            test_deactivation(
-                verificationStrategy = verificationStrategy,
-            )
-        }
-    }
 }
