+++ reactive/src/commonTest/kotlin/dev/toolkt/reactive/cell/Cell_map_misc_tests.kt
@@ -1,0 +1,49 @@
+package dev.toolkt.reactive.cell
+
+import dev.toolkt.reactive.cell.test_utils.CellVerificationStrategy
+import dev.toolkt.reactive.cell.test_utils.DynamicCellFactory
+import dev.toolkt.reactive.event_stream.EmitterEventStream
+import dev.toolkt.reactive.event_stream.map
+import kotlin.test.Test
+
+@Suppress("ClassName")
+class Cell_map_misc_tests {
+    private fun test_deactivation(
+        sourceCellFactory: DynamicCellFactory,
+        verificationStrategy: CellVerificationStrategy.Active,
+    ) {
+        val doTrigger = EmitterEventStream<Unit>()
+
+        val sourceCell = sourceCellFactory.createDynamicExternally(
+            initialValue = 10,
+            doUpdate = doTrigger.map { 11 },
+        )
+
+        val mapCell = sourceCell.map { it.toString() }
+
+        verificationStrategy.verifyDeactivation(
+            subjectCell = mapCell,
+            doTrigger = doTrigger,
+        )
+    }
+
+    private fun test_deactivation(
+        verificationStrategy: CellVerificationStrategy.Active,
+    ) {
+        DynamicCellFactory.values.forEach { sourceCellFactory ->
+            test_deactivation(
+                sourceCellFactory = sourceCellFactory,
+                verificationStrategy = verificationStrategy,
+            )
+        }
+    }
+
+    @Test
+    fun test_deactivation() {
+        CellVerificationStrategy.Active.values.forEach { verificationStrategy ->
+            test_deactivation(
+                verificationStrategy = verificationStrategy,
+            )
+        }
+    }
+}
